+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+- Unused Wally dependency Cmdr
+- Redundant Wally dependency TableUtil
+
 ## [1.1.0] - 2024-09-22
 
 ### Added