+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.1] - 2025-01-08
+
+### Added
+- Controllers can now specify a load order which affects which Start function runs first
 
 ### Removed
 - Unused Wally dependency Cmdr
@@ -102,7 +105,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - The base functionality of the entire framework!
 
-[unreleased]: https://github.com/StevenDahFish/fish/compare/v1.1.0...HEAD
+[1.1.1]: https://github.com/StevenDahFish/fish/compare/v1.1.0...HEAD
 [1.1.0]: https://github.com/StevenDahFish/fish/compare/v1.0.9...v1.1.0
 [1.0.9]: https://github.com/StevenDahFish/fish/compare/v1.0.8...v1.0.9
 [1.0.8]: https://github.com/StevenDahFish/fish/compare/v1.0.7...v1.0.8